requirements.txt
pytest
pytest-playwright
pytest-xdist
requests
allure-pytest
pytest-rerunfailures
jsonschema
psycopg2-binary